rename: chatmessage 도메인 이름 chat으로 변경
